+++ app/internal/factory/provideraccounts/options.go
@@ -14,12 +14,12 @@ import (
 type (
 	// Options - comment struct.
 	Options struct {
-		EventEmitter   mrsender.EventEmitter
-		UseCaseHelper  mrcore.UseCaseErrorWrapper
-		DBConnManager  mrstorage.DBConnManager
-		Locker         mrlock.Locker
-		RequestParsers RequestParsers
-		ResponseSender mrserver.ResponseSender
+		EventEmitter        mrsender.EventEmitter
+		UseCaseErrorWrapper mrcore.UseCaseErrorWrapper
+		DBConnManager       mrstorage.DBConnManager
+		Locker              mrlock.Locker
+		RequestParsers      RequestParsers
+		ResponseSender      mrserver.ResponseSender
 
 		UnitCompanyPage UnitCompanyPageOptions