+++ app/internal/factory/provideraccounts/section/pub/unit_company_page.go
@@ -25,7 +25,7 @@ func createUnitCompanyPage(ctx context.Context, opts provideraccounts.Options) (
 
 func newUnitCompanyPage(_ context.Context, opts provideraccounts.Options) (*httpv1.CompanyPage, error) { //nolint:unparam
 	storage := repository.NewCompanyPagePostgres(opts.DBConnManager)
-	useCase := usecase.NewCompanyPage(storage, opts.UseCaseHelper)
+	useCase := usecase.NewCompanyPage(storage, opts.UseCaseErrorWrapper)
 	controller := httpv1.NewCompanyPage(
 		opts.RequestParsers.ModuleParser,
 		opts.ResponseSender,